Allow Store::new and Store::new_with_path to return errors (resolves #19). Move Store code out of MT source and into a separate source file. Minor refactoring to help readability.
