+++ merkle/Cargo.toml
@@ -1,7 +1,9 @@
 [package]
 name = "merkletree"
-version = "0.10.0"
+version = "0.10.1"
 authors = [
+  "schomatis <lucas@protocol.ai>",
+  "nemo <nemo@protocol.ai>",
   "Ivan Prisyazhnyy <john.koepi@gmail.com>"
 ]
