+++ merkle/src/merkle.rs
@@ -1,15 +1,10 @@
 use hash::{Algorithm, Hashable};
-use memmap::MmapMut;
-use memmap::MmapOptions;
-use positioned_io::{ReadAt, WriteAt};
+use store::Store;
 use proof::Proof;
 use rayon::prelude::*;
-use std::fs::File;
 use std::iter::FromIterator;
 use std::marker::PhantomData;
-use std::ops::{self, Index};
 use std::sync::{Arc, RwLock};
-use tempfile::tempfile;
 
 /// Tree size (number of nodes) used as threshold to decide which build algorithm
 /// to use. Small trees (below this value) use the old build algorithm, optimized
@@ -91,406 +86,6 @@ pub trait Element: Ord + Clone + AsRef<[u8]> + Sync + Send + Default + std::fmt:
     fn copy_to_slice(&self, bytes: &mut [u8]);
 }
 
-/// Backing store of the merkle tree.
-pub trait Store<E: Element>:
-    ops::Deref<Target = [E]> + std::fmt::Debug + Clone + Send + Sync
-{
-    /// Creates a new store which can store up to `size` elements.
-    // FIXME: Return errors on failure instead of panicking
-    //  (see https://github.com/filecoin-project/merkle_light/issues/19).
-    fn new(size: usize) -> Self;
-
-    fn new_from_slice(size: usize, data: &[u8]) -> Self;
-
-    fn write_at(&mut self, el: E, i: usize);
-
-    // Used to reduce lock contention and do the `E` to `u8`
-    // conversion in `build` *outside* the lock.
-    // `buf` is a slice of converted `E`s and `start` is its
-    // position in `E` sizes (*not* in `u8`).
-    fn copy_from_slice(&mut self, buf: &[u8], start: usize);
-
-    fn read_at(&self, i: usize) -> E;
-    fn read_range(&self, r: ops::Range<usize>) -> Vec<E>;
-    fn read_into(&self, pos: usize, buf: &mut [u8]);
-
-    fn len(&self) -> usize;
-    fn is_empty(&self) -> bool;
-    fn push(&mut self, el: E);
-
-    // Sync contents to disk (if it exists). This function is used to avoid
-    // unnecessary flush calls at the cost of added code complexity.
-    fn sync(&self) {}
-}
-
-#[derive(Debug, Clone)]
-pub struct VecStore<E: Element>(Vec<E>);
-
-impl<E: Element> ops::Deref for VecStore<E> {
-    type Target = [E];
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<E: Element> Store<E> for VecStore<E> {
-    fn new(size: usize) -> Self {
-        VecStore(Vec::with_capacity(size))
-    }
-
-    fn write_at(&mut self, el: E, i: usize) {
-        if self.0.len() <= i {
-            self.0.resize(i + 1, E::default());
-        }
-
-        self.0[i] = el;
-    }
-
-    // NOTE: Performance regression. To conform with the current API we are
-    // unnecessarily converting to and from `&[u8]` in the `VecStore` which
-    // already stores `E` (in contrast with the `mmap` versions). We are
-    // prioritizing performance for the `mmap` case which will be used in
-    // production (`VecStore` is mainly for testing and backwards compatibility).
-    fn copy_from_slice(&mut self, buf: &[u8], start: usize) {
-        assert_eq!(buf.len() % E::byte_len(), 0);
-        let num_elem = buf.len() / E::byte_len();
-
-        if self.0.len() < start + num_elem {
-            self.0.resize(start + num_elem, E::default());
-        }
-
-        self.0.splice(
-            start..start + num_elem,
-            buf.chunks_exact(E::byte_len()).map(E::from_slice),
-        );
-    }
-
-    fn new_from_slice(size: usize, data: &[u8]) -> Self {
-        let mut v: Vec<_> = data
-            .chunks_exact(E::byte_len())
-            .map(E::from_slice)
-            .collect();
-        let additional = size - v.len();
-        v.reserve(additional);
-
-        VecStore(v)
-    }
-
-    fn read_at(&self, i: usize) -> E {
-        self.0[i].clone()
-    }
-
-    fn read_into(&self, i: usize, buf: &mut [u8]) {
-        self.0[i].copy_to_slice(buf);
-    }
-
-    fn read_range(&self, r: ops::Range<usize>) -> Vec<E> {
-        self.0.index(r).to_vec()
-    }
-
-    fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-
-    fn push(&mut self, el: E) {
-        self.0.push(el);
-    }
-}
-
-#[derive(Debug)]
-pub struct MmapStore<E: Element> {
-    store: MmapMut,
-    len: usize,
-    _e: PhantomData<E>,
-}
-
-impl<E: Element> ops::Deref for MmapStore<E> {
-    type Target = [E];
-
-    fn deref(&self) -> &Self::Target {
-        unimplemented!()
-    }
-}
-
-impl<E: Element> Store<E> for MmapStore<E> {
-    #[allow(unsafe_code)]
-    fn new(size: usize) -> Self {
-        let byte_len = E::byte_len() * size;
-
-        MmapStore {
-            store: MmapOptions::new().len(byte_len).map_anon().unwrap(),
-            len: 0,
-            _e: Default::default(),
-        }
-    }
-
-    fn new_from_slice(size: usize, data: &[u8]) -> Self {
-        assert_eq!(data.len() % E::byte_len(), 0);
-
-        let mut res = Self::new(size);
-
-        let end = data.len();
-        res.store[..end].copy_from_slice(data);
-        res.len = data.len() / E::byte_len();
-
-        res
-    }
-
-    // Writing at positions `i` will mark all other positions as
-    // occupied with respect to the `len()` so the new `len()`
-    // is `>= i`.
-    fn write_at(&mut self, el: E, i: usize) {
-        let b = E::byte_len();
-        self.store[i * b..(i + 1) * b].copy_from_slice(el.as_ref());
-        self.len = std::cmp::max(self.len, i + 1);
-    }
-
-    fn copy_from_slice(&mut self, buf: &[u8], start: usize) {
-        let b = E::byte_len();
-        assert_eq!(buf.len() % b, 0);
-        self.store[start * b..start * b + buf.len()].copy_from_slice(buf);
-        self.len = std::cmp::max(self.len, start + buf.len() / b);
-    }
-
-    fn read_at(&self, i: usize) -> E {
-        let b = E::byte_len();
-        let start = i * b;
-        let end = (i + 1) * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        E::from_slice(&self.store[start..end])
-    }
-
-    fn read_into(&self, i: usize, buf: &mut [u8]) {
-        let b = E::byte_len();
-        let start = i * b;
-        let end = (i + 1) * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        buf.copy_from_slice(&self.store[start..end]);
-    }
-
-    fn read_range(&self, r: ops::Range<usize>) -> Vec<E> {
-        let b = E::byte_len();
-        let start = r.start * b;
-        let end = r.end * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        self.store[start..end]
-            .chunks(b)
-            .map(E::from_slice)
-            .collect()
-    }
-
-    fn len(&self) -> usize {
-        self.len
-    }
-
-    fn is_empty(&self) -> bool {
-        self.len == 0
-    }
-
-    fn push(&mut self, el: E) {
-        let l = self.len;
-        assert!(
-            (l + 1) * E::byte_len() <= self.store.len(),
-            "not enough space"
-        );
-
-        self.write_at(el, l);
-    }
-}
-
-impl<E: Element> Clone for MmapStore<E> {
-    fn clone(&self) -> MmapStore<E> {
-        MmapStore::new_from_slice(
-            self.store.len() / E::byte_len(),
-            &self.store[..(self.len() * E::byte_len())],
-        )
-    }
-}
-
-/// Disk-only store use to reduce memory to the minimum at the cost of build
-/// time performance. Most of its I/O logic is in the `store_copy_from_slice`
-/// and `store_read_range` functions.
-#[derive(Debug)]
-pub struct DiskStore<E: Element> {
-    len: usize,
-    _e: PhantomData<E>,
-    file: File,
-
-    // We cache the `store.len()` call to avoid accessing disk unnecessarily.
-    // Not to be confused with `len`, this saves the total size of the `store`
-    // in bytes and the other one keeps track of used `E` slots in the `DiskStore`.
-    store_size: usize,
-}
-
-impl<E: Element> ops::Deref for DiskStore<E> {
-    type Target = [E];
-
-    fn deref(&self) -> &Self::Target {
-        unimplemented!()
-    }
-}
-
-impl<E: Element> Store<E> for DiskStore<E> {
-    #[allow(unsafe_code)]
-    fn new(size: usize) -> Self {
-        let byte_len = E::byte_len() * size;
-        let file = tempfile().expect("couldn't create temp file");
-        file.set_len(byte_len as u64)
-            .unwrap_or_else(|_| panic!("couldn't set len of {}", byte_len));
-
-        DiskStore {
-            len: 0,
-            _e: Default::default(),
-            file,
-            store_size: byte_len,
-        }
-    }
-
-    fn new_from_slice(size: usize, data: &[u8]) -> Self {
-        assert_eq!(data.len() % E::byte_len(), 0);
-
-        let mut res = Self::new(size);
-
-        res.store_copy_from_slice(0, data);
-        res.len = data.len() / E::byte_len();
-
-        res
-    }
-
-    fn write_at(&mut self, el: E, i: usize) {
-        let b = E::byte_len();
-        self.store_copy_from_slice(i * b, el.as_ref());
-        self.len = std::cmp::max(self.len, i + 1);
-    }
-
-    fn copy_from_slice(&mut self, buf: &[u8], start: usize) {
-        let b = E::byte_len();
-        assert_eq!(buf.len() % b, 0);
-        self.store_copy_from_slice(start * b, buf);
-        self.len = std::cmp::max(self.len, start + buf.len() / b);
-    }
-
-    fn read_at(&self, i: usize) -> E {
-        let b = E::byte_len();
-        let start = i * b;
-        let end = (i + 1) * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        E::from_slice(&self.store_read_range(start, end))
-    }
-
-    fn read_into(&self, i: usize, buf: &mut [u8]) {
-        let b = E::byte_len();
-        let start = i * b;
-        let end = (i + 1) * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        self.store_read_into(start, end, buf);
-    }
-
-    fn read_range(&self, r: ops::Range<usize>) -> Vec<E> {
-        let b = E::byte_len();
-        let start = r.start * b;
-        let end = r.end * b;
-        let len = self.len * b;
-        assert!(start < len, "start out of range {} >= {}", start, len);
-        assert!(end <= len, "end out of range {} > {}", end, len);
-
-        self.store_read_range(start, end)
-            .chunks(b)
-            .map(E::from_slice)
-            .collect()
-    }
-
-    fn len(&self) -> usize {
-        self.len
-    }
-
-    fn is_empty(&self) -> bool {
-        self.len == 0
-    }
-
-    fn push(&mut self, el: E) {
-        let l = self.len;
-        assert!(
-            (l + 1) * E::byte_len() <= self.store_size(),
-            format!(
-                "not enough space, l: {}, E size {}, store len {}",
-                l,
-                E::byte_len(),
-                self.store_size()
-            )
-        );
-
-        self.write_at(el, l);
-    }
-
-    fn sync(&self) {
-        self.file.sync_all().expect("failed to sync file");
-    }
-}
-
-impl<E: Element> DiskStore<E> {
-    pub fn store_size(&self) -> usize {
-        self.store_size
-    }
-
-    pub fn store_read_range(&self, start: usize, end: usize) -> Vec<u8> {
-        let read_len = end - start;
-        let mut read_data = vec![0; read_len];
-
-        assert_eq!(
-            self.file
-                .read_at(start as u64, &mut read_data)
-                .unwrap_or_else(|_| panic!(
-                    "failed to read {} bytes from file at offset {}",
-                    read_len, start
-                )),
-            read_len
-        );
-
-        read_data
-    }
-
-    pub fn store_read_into(&self, start: usize, end: usize, buf: &mut [u8]) {
-        buf.copy_from_slice(&self.store_read_range(start, end));
-    }
-
-    pub fn store_copy_from_slice(&mut self, start: usize, slice: &[u8]) {
-        assert!(start + slice.len() <= self.store_size);
-        self.file
-            .write_at(start as u64, slice)
-            .expect("failed to write file");
-    }
-}
-
-// FIXME: Fake `Clone` implementation to accommodate the artificial call in
-//  `from_data_with_store`, we won't actually duplicate the mmap memory,
-//  just recreate the same object (as the original will be dropped).
-impl<E: Element> Clone for DiskStore<E> {
-    fn clone(&self) -> DiskStore<E> {
-        unimplemented!("We can't clone a store with an already associated file");
-    }
-}
-
 impl<T: Element, A: Algorithm<T>, K: Store<T>> MerkleTree<T, A, K> {
     /// Creates new merkle from a sequence of hashes.
     pub fn new<I: IntoIterator<Item = T>>(data: I) -> MerkleTree<T, A, K> {
@@ -518,8 +113,7 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> MerkleTree<T, A, K> {
     //  store adding a `capacity` method to the trait.
     pub fn from_leaves_store(leaves: K, leafs: usize) -> MerkleTree<T, A, K> {
         let pow = next_pow2(leafs);
-
-        let top_half = K::new(pow);
+        let top_half = K::new(pow).expect("Failed to create top_half");
 
         Self::build(leaves, top_half, leafs, log2_pow2(2 * pow))
     }
@@ -528,7 +122,6 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> MerkleTree<T, A, K> {
     fn build(leaves: K, top_half: K, leafs: usize, height: usize) -> Self {
         // This algorithms assumes that the underlying store has preallocated enough space.
         // TODO: add an assert here to ensure this is the case.
-
         if leafs <= SMALL_TREE_BUILD {
             return Self::build_small_tree(leaves, top_half, leafs, height);
         }
@@ -651,10 +244,7 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> MerkleTree<T, A, K> {
 
         MerkleTree {
             leaves: Arc::try_unwrap(leaves_lock).unwrap().into_inner().unwrap(),
-            top_half: Arc::try_unwrap(top_half_lock)
-                .unwrap()
-                .into_inner()
-                .unwrap(),
+            top_half: Arc::try_unwrap(top_half_lock).unwrap().into_inner().unwrap(),
             leafs,
             height,
             root,
@@ -865,8 +455,8 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> MerkleTree<T, A, K> {
         let leafs_count = leafs.len() / T::byte_len();
         let pow = next_pow2(leafs_count);
 
-        let leaves = K::new_from_slice(pow, leafs);
-        let top_half = K::new(pow);
+        let leaves = K::new_from_slice(pow, leafs).expect("Failed to create leaves");
+        let top_half = K::new(pow).expect("Failed to create top_half");
 
         assert!(leafs_count > 1);
         Self::build(leaves, top_half, leafs_count, log2_pow2(2 * pow))
@@ -897,8 +487,8 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> FromIndexedParallelIterator<T>
         let leafs = iter.opt_len().expect("must be sized");
         let pow = next_pow2(leafs);
 
-        let mut leaves = K::new(pow);
-        let top_half = K::new(pow);
+        let mut leaves = K::new(pow).expect("Failed to create leaves");
+        let top_half = K::new(pow).expect("Failed to create top_half");
 
         populate_leaves_par::<T, A, K, _>(&mut leaves, iter);
 
@@ -916,8 +506,8 @@ impl<T: Element, A: Algorithm<T>, K: Store<T>> FromIterator<T> for MerkleTree<T,
 
         let pow = next_pow2(leafs);
 
-        let mut leaves = K::new(pow);
-        let top_half = K::new(pow);
+        let mut leaves = K::new(pow).expect("Failed to create leaves");
+        let top_half = K::new(pow).expect("Failed to create top_half");
 
         populate_leaves::<T, A, K, I>(&mut leaves, iter);
