+++ merkle/src/test_sip.rs
@@ -1,9 +1,10 @@
 #![cfg(test)]
 
 use hash::{Algorithm, Hashable};
+use store::VecStore;
 use merkle::log2_pow2;
 use merkle::next_pow2;
-use merkle::{MerkleTree, VecStore};
+use merkle::MerkleTree;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::Hasher;
 use std::iter::FromIterator;
